Fix RSSI display value
The RSSI display value is wrong. -23dBm displays as S9+70dB, the correct display should be S9+50dB since S9 is equal to 73dBm (https://en.wikipedia.org/wiki/S_meter). The display error is caused by line 531: 
$relint = intval($dBraw) + 93;
This code change substitutes 93 with 73, which is the correct dBm for S9.