Add support for CSS dark mode.
This patch adds a new stylesheet, pistar-dark.css, and links it in
all of the relevant PHP files.

This utilizes the (prefers-color-scheme: dark) CSS Media Query, as
defined in Level 5.  If the user agent says that it prefers darker
color schemes, then the pistar-dark.css stylesheet will be applied
to the page dynamically.

We could not use @import to add this to the pistar-css*.php files,
as those settings would override normal selectors if we placed the
directive at the top, and it would be ignored at the bottom of the
file.  Thus, adding a second <link> tag to each and every page was
the best course of action.