Add support for CSS dark mode (editable).
This adds support for the Level 5 prefers-color-scheme CSS media query
in order to automatically apply a dark theme if the user's browser
says it prefers one.  Colors are generated from values specified in
the pistar-css.ini file, and falling back to defaults if they are not
defined there (which will be the case for a majority of users).

In contrast to my previous versions, the two main stylesheets are now
dynamically updated with dark mode support, should the user not opt
out.  Whether we make this opt-in or opt-out is a decision best left
to the project team. :-)