Update version.php
Better fix for the issue seen on some D-Star reflectors where the end of network transmission data is missing, and instead the end of the tranmission ends with "starting fast data mode" - this will now show a "TOut" / time-out rather than the proper end of call (because there isn't one). I don't expect side effects from this, but we'll keep an eye on it, at least it does not show the repeater stuck in TX (as it did previously).